LILY-1168: increase max street number

+++ lily/utils/models/models.py
@@ -337,7 +337,7 @@ class Address(TenantMixin):
     )
 
     street = models.CharField(max_length=255, verbose_name=_('street'), blank=True)
-    street_number = models.SmallIntegerField(verbose_name=_('street number'), blank=True, null=True)
+    street_number = models.IntegerField(verbose_name=_('street number'), blank=True, null=True)
     complement = models.CharField(max_length=255, verbose_name=_('complement'), blank=True, null=True)
     postal_code = models.CharField(max_length=10, verbose_name=_('postal code'), blank=True)
     city = models.CharField(max_length=100, verbose_name=_('city'), blank=True)
